+++ create-script/CHANGELOG.md
@@ -12,6 +12,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   `Fixed` for any bug fixes
 -   `Security` in case of vulnerabilities
 
+## [UNRELEASED]
+
+### Removed
+
+-   Standard software license
+
 ## [0.1.7] - 2025-11-17
 
 ### Changed